refactor: 소셜로그인 상수 분리
