+++ src/constants/socialLogin.tsx
@@ -1,0 +1,13 @@
+import GoogleIcon from "../assets/icon/google-round.svg?react";
+import KaKaoIcon from "../assets/icon/kakao-round.svg?react";
+
+const SOCIAL_LOGIN = [
+  { name: "구글", href: "https://www.google.com/", icon: <GoogleIcon /> },
+  {
+    name: "카카오",
+    href: "https://www.kakaocorp.com/page/",
+    icon: <KaKaoIcon />,
+  },
+];
+
+export default SOCIAL_LOGIN;
